requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.36
python-dotenv==1.0.1
Flask-RESTful==0.3.10
marshmallow==3.21.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0

